test(protocol): port account registry relayer functionality tests over
